+++ CHANGELOG.md
@@ -17,10 +17,13 @@ Updated README.md
 Updated README.md
 
 ## 0.0.6
-Update main class from WhatsAppSender to WhatsAppSenderFlutter
+Updated main class from WhatsAppSender to WhatsAppSenderFlutter
 
 ## 0.0.7
 Updated README.md
 
 ## 0.0.8
 Fixed example
+
+## 0.0.9
+Added multi session support